strawberry-django/pyproject.toml
[project]
name = "imfaxblog"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yasuhiro TODOROKI", email = "yasu.todo@gmail.com" }
]
dependencies = [
    "django>=5.1.7",
    "strawberry-graphql>=0.262.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/imfaxblog"]
